Changed the dates format of the build JSON files
The previous format of the dates associated to the parameters "committed_at", "started_at", and "finished_at" was yyyy-MM-dd'T'HH:mm:ss'Z'.
The new format of the dates associated to the parameters "committed_at", "started_at", and "finished_at" is yyyy-MM-dd'T'HH:mm:ss.SSS'Z'.

Thus, the new dates formate contains also the milliseconds.

Here is an example:

2017-02-01T17:11:09Z becomes 2017-02-01T17:11:09.000Z